update broken userchrome for v75
Around FF v70, everything broke again and I have been unable to get most of my modifications to work correctly. The current version still moves the menu button to the top left, but not next to the tabs anymore--I couldn't get this to work correctly.

On the plus side, the new CSS is a lot simpler.